HOTFIX and Mapping Deployment 8th March 2025

+++ src/main/java/com/clm/matching/models/CategoryMatchResponseDTO.java
@@ -13,6 +13,6 @@ import java.util.List;
 public class CategoryMatchResponseDTO {
 
     private Long categoryId;
-    private double matchPercentage;
+    private int matchPercentage;
     private List<Long> matchedOptions;
 }
